Move send methods from abstract object to dispatch code
Really, the methods don’t belong into the objects. They belong into the execution code.

And avoids having interpreter-specific code in the objects.
