+++ src/som/interpreter/ast/nodes/global_read_node.py
@@ -1,3 +1,4 @@
+from .dispatch import lookup_and_send
 from .expression_node import ExpressionNode
 from som.vm.globals import nilObject, trueObject, falseObject
 
@@ -15,8 +16,13 @@ class UninitializedGlobalReadNode(ExpressionNode):
         if self._universe.has_global(self._global_name):
             return self._specialize().execute(frame)
         else:
-            return frame.get_self().send_unknown_global(self._global_name,
-                                                        self._universe)
+            return self.send_unknown_global(
+                frame.get_self(), self._global_name, self._universe)
+
+    @staticmethod
+    def send_unknown_global(receiver, global_name, universe):
+        arguments = [global_name]
+        return lookup_and_send(receiver, "unknownGlobal:", arguments, universe)
 
     def _specialize(self):
         glob = self._global_name.get_embedded_string()
